+++ searchrequest/go.mod
@@ -1,0 +1,3 @@
+module github.com/Nickhil-Sethi/gogrep/searchrequest
+
+go 1.12
